# src/Hedwig/ClientApp/src/OAS-generated/models/Age.ts
// tslint:disable
// eslint-disable
/**
 * Hedwig API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum Age {
    Infant = 'Infant',
    Preschool = 'Preschool',
    School = 'School'
}

export function AgeFromJSON(json: any): Age {
    return AgeFromJSONTyped(json, false);
}

export function AgeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Age {
    return json as Age;
}

export function AgeToJSON(value?: Age | null): any {
    return value as any;
}