# src/Hedwig/ClientApp/src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hedwig API",
    "version": "v1"
  },
  "paths": {
    "/api/organizations/{orgId}/Children": {
      "get": {
        "tags": [
          "Hedwig"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include[]",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Child"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Child"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Child"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Hedwig"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{orgId}/Children/{id}": {
      "get": {
        "tags": [
          "Hedwig"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include[]",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hedwig"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Child"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/organizations/{orgId}/Sites": {
      "get": {
        "tags": [
          "Hedwig"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/organizations/{orgId}/Sites/{id}": {
      "get": {
        "tags": [
          "Hedwig"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include[]",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Age": {
        "enum": [
          "Infant",
          "Preschool",
          "School"
        ],
        "type": "string"
      },
      "FundingSource": {
        "enum": [
          "CDC",
          "C4K"
        ],
        "type": "string"
      },
      "FundingTime": {
        "enum": [
          "Full",
          "Part"
        ],
        "type": "string"
      },
      "OrganizationPermission": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "integer",
            "format": "int32"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "SitePermission": {
        "type": "object",
        "properties": {
          "siteId": {
            "type": "integer",
            "format": "int32"
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "User": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "wingedKeysId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "suffix": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "orgPermissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationPermission"
            },
            "nullable": true
          },
          "sitePermissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SitePermission"
            },
            "nullable": true
          },
          "sites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Funding": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "enrollmentId": {
            "type": "integer",
            "format": "int32"
          },
          "enrollment": {
            "$ref": "#/components/schemas/Enrollment"
          },
          "source": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "time": {
            "$ref": "#/components/schemas/FundingTime"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "Enrollment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "child": {
            "$ref": "#/components/schemas/Child"
          },
          "siteId": {
            "type": "integer",
            "format": "int32"
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          },
          "age": {
            "$ref": "#/components/schemas/Age"
          },
          "entry": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "exit": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fundings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Funding"
            },
            "nullable": true
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "Site": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "organizationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "enrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReportingPeriod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "period": {
            "type": "string",
            "format": "date-time"
          },
          "periodStart": {
            "type": "string",
            "format": "date-time"
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "OrganizationReport": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "integer",
            "format": "int32"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/FundingSource"
          },
          "reportingPeriodId": {
            "type": "integer",
            "format": "int32"
          },
          "reportingPeriod": {
            "$ref": "#/components/schemas/ReportingPeriod"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Organization": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "sites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "nullable": true
          },
          "reports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationReport"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Gender": {
        "enum": [
          "Male",
          "Female",
          "Unknown",
          "Unspecified"
        ],
        "type": "string"
      },
      "FamilyDetermination": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfPeople": {
            "type": "integer",
            "format": "int32"
          },
          "income": {
            "type": "number",
            "format": "double"
          },
          "determined": {
            "type": "string",
            "format": "date-time"
          },
          "familyId": {
            "type": "integer",
            "format": "int32"
          },
          "family": {
            "$ref": "#/components/schemas/Family"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "Family": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Child"
            },
            "nullable": true
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "town": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "string",
            "nullable": true
          },
          "homelessness": {
            "type": "boolean"
          },
          "determinations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyDetermination"
            },
            "nullable": true
          },
          "organizationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "Child": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sasid": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "suffix": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "birthdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "birthTown": {
            "type": "string",
            "nullable": true
          },
          "birthState": {
            "type": "string",
            "nullable": true
          },
          "birthCertificateId": {
            "type": "string",
            "nullable": true
          },
          "americanIndianOrAlaskaNative": {
            "type": "boolean"
          },
          "asian": {
            "type": "boolean"
          },
          "blackOrAfricanAmerican": {
            "type": "boolean"
          },
          "nativeHawaiianOrPacificIslander": {
            "type": "boolean"
          },
          "white": {
            "type": "boolean"
          },
          "hispanicOrLatinxEthnicity": {
            "type": "boolean"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "foster": {
            "type": "boolean"
          },
          "familyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "family": {
            "$ref": "#/components/schemas/Family"
          },
          "enrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Enrollment"
            },
            "nullable": true
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": false
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}